Crate irq_safety

Types for interrupt-safe operations, i.e., avoiding deadlock when sharing data between a regular thread context and an interrupt handler context.

Key types include:

HeldInterrupts: a guard type that auto-reenables interrupts when dropped, only if they were originally enabled when the guard was created.

[ MutexIrqSafe] and [RwLockIrqSafe]: spinlock wrappers that use [spin::Mutex] and [spin::RwLock`] internally to auto-disable interrupts for the duration of the lock being held.

Structs

HeldInterrupts

A handle for frozen interrupts

MutexIrqSafe

This type provides interrupt-safe MUTual EXclusion based on [spin::Mutex].

MutexIrqSafeGuard

A guard to which the protected data can be accessed

RwLockIrqSafe

A simple wrapper around a RwLock whose guards disable interrupts properly

RwLockIrqSafeReadGuard

A guard to which the protected data can be read

RwLockIrqSafeWriteGuard

A guard to which the protected data can be written

Functions

disable_interrupts

enable_interrupts

hold_interrupts

Prevent interrupts from firing until the return value is dropped (goes out of scope). After it is dropped, the interrupts are returned to their prior state, not blindly re-enabled.

interrupts_enabled